pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask-cors>=6.0.1",
    "flask>=3.1.1",
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
]
